# docker/api-server/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "This is a sample server Todo server."
  version: "1.0.0"
  title: "Swagger Todo"
  termsOfService: "http://swagger.io/terms/"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "127.0.0.1:8083"
basePath: "/v2"
tags:
- name: "todo"
schemes:
  - "http"
paths:
  /todo/:
    get:
      tags:
      - "todo"
      summary: "todoアイテム一覧を取得する"
      operationId: "todoGET"
      parameters: []
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/inline_response_200"
        "404":
          description: "Error"
        "500":
          description: "Error"
      x-swagger-router-controller: "Todo"
    post:
      tags:
      - "todo"
      summary: "todoアイテムを新規作成する"
      operationId: "todoPOST"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/body"
      responses:
        "200":
          description: "Success"
        "500":
          description: "Error"
      x-swagger-router-controller: "Todo"
  /todo/{id}/:
    put:
      tags:
      - "todo"
      summary: "todoアイテムを更新する"
      operationId: "todoIdPUT"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/body_1"
      responses:
        "200":
          description: "Success"
        "404":
          description: "Error"
        "500":
          description: "Error"
      x-swagger-router-controller: "Todo"
    delete:
      tags:
      - "todo"
      summary: "todoアイテムを削除する"
      operationId: "todoIdDELETE"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Success"
        "404":
          description: "Error"
        "500":
          description: "Error"
      x-swagger-router-controller: "Todo"
definitions:
  body_1:
    type: "object"
    properties:
      status:
        type: "string"
        example: "waiting"
      title:
        type: "string"
        example: "皿を洗う"
      text:
        type: "string"
        example: "10時までに皿を洗う"
  inline_response_200:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      status:
        type: "string"
        example: "waiting"
      title:
        type: "string"
        example: "皿を洗う"
      text:
        type: "string"
        example: "10時までに皿を洗う"
    example:
      id: 0
      text: "10時までに皿を洗う"
      title: "皿を洗う"
      status: "waiting"
  body:
    type: "object"
    properties:
      title:
        type: "string"
        example: "皿を洗う"
      text:
        type: "string"
        example: "10時までに皿を洗う"